Break out `up`, `destroy`, and other operations
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out
operational commands such as `up`, `destroy`, `refresh`, `import` etc.
